requirements.txt
langchain
openai
faiss-cpu
requests
pydantic
fastapi
uvicorn
docker
crewai
pyautogen
PyYAML